ceph: separate the creation of Mon PVs and OSD PV in the test
OSDs are backed by many types of devices (ex. lvm and mpath) in OSD on PVC.
When we add the tests of these types later, PVs for OSDs should be deleted
before testing the next types to reuse these volumes. On the other hand,
Mons will keep existing until the end of the test suite.

Let's separate the creation of both MONs and OSDs to ease the implementation
of the above-mentioned tests

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>